# apps/leaderboard-client/src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Leaderboard API
  description: |
    API REST du projet Leaderboard — gestion de projets, challenges, tâches, contributions, évaluations et sync meetings.

    ## Authentification
    L'API utilise des cookies HTTP-only (`access_token` / `refresh_token`) via OAuth2 Google.
    Certaines routes nécessitent le rôle **admin**.
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API routes
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    cronBearer:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        details:
          type: array
          items:
            type: object
      required: [error, details]
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true

    # ── Users ──
...
    Task:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        challenge_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
          description: Owner of this task's personal board. Null means this is a template task (edited by the challenge admin/manager and copied onto each contributor's board on join).
        parent_task_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        created_at:
          type: string
          format: date-time
...
    UpdateTask:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        parent_task_id:
          type: string
          format: uuid
          nullable: true
...
paths:
...
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Tasks]
      summary: Récupérer une tâche
      responses:
        '200':
          description: Tâche trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Tâche non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Tasks]
      summary: Mettre à jour une tâche
      description: Réservé au propriétaire de la tâche (ou admin/manager du challenge pour une tâche modèle).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
      responses:
        '200':
          description: Tâche mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
    delete:
      tags: [Tasks]
      summary: Supprimer une tâche
      responses:
        '200':
          description: Tâche supprimée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
...